pyproject.toml
[tool.poetry]
name = "ml-trading-bot"
version = "0.1.0"
description = "Daily adaptive ML trading bot for 1H/4H"
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
numpy = "^1.26.0"
polars = "^1.4.0"
scikit-learn = "^1.5.0"
lightgbm = "^4.3.0"
xgboost = "^2.0.0"
ta = "^0.11.0"
mlflow = "^2.12.0"
prefect = "^3.0.0"
pydantic = "^2.7.0"
ccxt = "^4.3.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.3.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Added GUI/API deps
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
streamlit = "^1.36.0"
requests = "^2.32.0"
